sched: task structure needs to define a guest state
This guest state structure will contain all the information about the VM -
the address space, the SIE control block, VM register save area, etc. To
save some space, this guest state struct is pointed to by a field in the
tast struct - this way all the nucleus service tasks do not have to allocate
(and not use) all these fields.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>