io: use a monotonic counter as the IO parameter
Since the control units use the IO parameter value until it gets changed
with either a MSCH or SSCH, it is necessary to use a monotonically
increasing variable to distinguish between CPU-initiated IO operation and a
device-initiated IO operation.

This patch removes the rather crufty ops array, and uses a much cleaner and
more natural linked list.

1) CPU initiated operation:
      There is an io_op struct in the operations list that matches the
      <interruption parameter, subchannel> tuple. Once the structure is
      found, proceed as previously - TSCH, etc.

2) Device initiated operation
      If the <param, subchannel> tuple was not found in the ops list, the
      interruption must have been initiated by the device. The driver needs
      to be informed about the condition.

      NOTE: The driver callback is not yet implemented.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>