cp: revamp of command parsing
The new code allows for arbitrary nesting (stack permitting) of CP
sub-commands. It also implements a DISPLAY STORAGE stub that should be
rather trivial to fill in.

Note: This code still has a problem with ambiguity of abbreviations. If
there are two valid commands (e.g., FOOBAR and FOOFOO), and the user tries
to run FOO, the first command encountered during the check (FOOBAR or
FOOFOO) will get executed.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>