cp: logic to issue SIE
If the guest is in the running state, SIE will be issued in a loop.

Note 1: I _think_ that the SIE control block isn't set up properly because
SIE doesn't appear to change the guest state (the guest regs, etc.).

Note 2: This commit does NOT handle interceptions, and interrupt delivery.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>