sched: interrupted SIE workaround
If an interruption occurs during the execution of SIE, the old PSW's address
will point to the SIE instruction and NOT to the following instruction. This
tiny workaround in the scheduler effectively skips the SIE instruction. Once
there is a proper way to send interruption signals to SIE, this can get
removed.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>