cp: guest IPL helper code (incomplete)
The Principles of Operation say that the IPL IO happens as if...

	The read operation is performed as if a START SUBCHANNEL instruction
	were executed that designated (1) the subchannel corresponding to
	the device number specified by the load-unit-address controls and
	(2) an ORB containing all zeros, except for a byte of all ones in
	the logical-path-mask field. (SA22-7832-06, page 17-17)

This helper will do exactly that. So, instead of the hypervisor
special-casing the IPL operation, it'll simply copy this .text into the
guest's address space (somewhere above the 16MB mark), modify the PSW, and
begin execution of this code.

The helper will simply issue the IO as the POP states. The hypervisor will
intercept all IO operations as during normal execution. Once the IPL related
IO is complete, the helper signals the hypervisor that is it done. The
hypervisor wipes the helper from the guest storage (memset if the contaning
the helper is supposed to be in the configuration, or unmapping of the page
otherwise).

Note: The above description is the design idea, the code in this commit is
NOT yet complete.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>