cp: clean up guest_state and directory handling
The directory and guest_state structures got abused a several ways. This
commit removes that abuse by doing a major refactoring of the code. Now, the
top-level structure that defines a virtual system, is struct virt_sys.

struct virt_sys refers to the console used to communicate with the user,
contains several system parameters (e.g., guest storage size), points to the
entry in the directory, and points to the struct task which is responsible
for emulating the cpu.

This is a much more logical structure, and will make SMP guests much easier
to do.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>