cp: rewrite of DISPLAY STORAGE
DISPLAY STORAGE now accepts an option length argument, and the formating
character can be one of: 'N' (numeric - hex display), 'I' (disassembled
instructions).  Numeric display is the default.

The length is specified in hex. For numeric display, it is rounded up to the
nearest word size. For disassembly, the all instructions _beginning_ within
the specified byte range are displayed.

Examples:

	DISPLAY STORAGE 12345
	DISPLAY STORAGE 12345.200
	DISPLAY STORAGE N12345
	DISPLAY STORAGE N12345.200
	DISPLAY STORAGE I12345
	DISPLAY STORAGE I12345.200

The first pair, and second pair of commands is identical. The second,
fourth, and sixth commands display at least 0x200 bytes.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>