console: change the maximum number of console lines to flush
The previous maximum (8) was large enough to test both cases - merged and
non-merged, however as there are numerous ways to cause a lot of console IO,
up the limit to something more reasonable (e.g., 32). This also means, that
we can send over 4kB to the console with one IO (32 lines/IO * 150
chars/line).

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>