dat: reset all entries to 0xff when allocating region/segment/page tables
The code had a bug where it reset only the first page of entries, so if one
attempted to translate an address that was high enough to use the 2nd-4th
page it would get an entry full of zeros.  This translated as origin 0 &
invalid 0.

Thanks goes to Ivan Warren for finding this bug.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>