cp: make DISPLAY PSW and STORE PSW commands mode-aware
Before this commit, DISPLAY PSW always displayed 4 words. This was
misleading as there is no indication which architecture mode the guest is
in.

Similarly, STORE PSW always accepted up to 4 words, and stored them to
potentially unused area of the SIE control block (the 8 bytes not used by
ESA/390 PSWs).

Closes bug #134.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>