cp: rename the 31bit field in the psw
It turns out that in ESA/390 mode, the instruction address is stored in the
regular 31bit field (not the same place as the z/Arch instruction address).
Let's give it a friendly name.

Additionally, let's label bit 12 as the one that gets set to either 1 or 0
depending on the PSW format.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>