cp: replace crusty console line buffering mechanism with a spoolfile
The spool file layer needs to exist anyways so let's use it.  It does mean
that a console I/O from a guest will end up being copied 3 times (guest
buffer -> spooldev_cons buffer -> spoolfile -> I/O buffer).  The is no way
to avoid at least one copy.  At this point, it's pointless to try to
optimize that and so we'll live this behavior for now.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>