Convert non-tools build system to plain old makefiles (bmake-flavor)
CMake works great when not cross compiling, and that's why we continue using
it for tools building.  When it comes to cross-compling, it is much simpler
to just use a hand-crafted makefile that does all the heavy lifting.

This build-a-stand-alone-binary makefile still has some rough edges, but it
seems to work well enough.

bmake seems to be the simplest variant of make to accomplish all this
without too much pain.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>